#27 Split Spring configs up more
