+++ rabbitmq-operator/src/main/java/com/indeed/operators/rabbitmq/config/AppConfig.java
@@ -54,59 +54,8 @@ public class AppConfig {
     private static final int SCHEDULED_THREAD_POOL_SIZE = 4;
 
     @Bean
-    public NetworkPartitionReconciler networkPartitionReconciler(
-            final RabbitMQResourceController rabbitMQResourceController,
-            final NetworkPartitionResourceController networkPartitionResourceController,
-            final AreQueuesEmptyOperation queuesEmptyOperation,
-            final RabbitMQPods rabbitMQPods,
-            final RabbitMQContainers rabbitMQContainers,
-            final StatefulSetController statefulSetController,
-            final PodController podController,
-            final String namespace
-    ) {
-        return new NetworkPartitionReconciler(rabbitMQResourceController, networkPartitionResourceController, queuesEmptyOperation, rabbitMQPods, rabbitMQContainers, statefulSetController, podController, namespace);
-    }
-
-    @Bean
     public KubernetesClient kubernetesClient() {
         return new DefaultKubernetesClient();
-    }
-
-    @Bean
-    public AreQueuesEmptyOperation queuesEmptyOperation(
-            final RabbitMQApiClient rabbitMQApiClient
-    ) {
-        return new AreQueuesEmptyOperation(rabbitMQApiClient);
-    }
-
-    @Bean
-    public RabbitMQPods rabbitMQPods() {
-        return new RabbitMQPods();
-    }
-
-    @Bean
-    public RabbitMQContainers rabbitMQContainers() {
-        return new RabbitMQContainers();
-    }
-
-    @Bean
-    public RabbitMQSecrets rabbitSecrets(
-            final Function<Integer, String> randomStringGenerator
-    ) {
-        return new RabbitMQSecrets(randomStringGenerator, DEFAULT_USERNAME);
-    }
-
-    @Bean
-    public RabbitMQServices rabbitMQServices() {
-        return new RabbitMQServices();
-    }
-
-    @Bean
-    public RabbitMQApiClient rabbitMQApiClient(
-            final OkHttpClient okHttpClient,
-            final SecretsController secretsController
-    ) {
-        return new RabbitMQApiClient(okHttpClient, new ObjectMapper(), secretsController);
     }
 
     @Bean
@@ -115,70 +64,8 @@ public class AppConfig {
     }
 
     @Bean
-    public RabbitMQEventWatcher rabbitEventWatcher(
-            final RabbitMQClusterReconciler reconciler,
-            final RabbitMQResourceController controller,
-            final ClusterReconciliationOrchestrator orchestrator
-    ) {
-        return new RabbitMQEventWatcher(reconciler, controller, orchestrator);
-    }
-
-    @Bean
     public NamedSemaphores namedSemaphores() {
         return new NamedSemaphores();
-    }
-
-    @Bean
-    public RabbitMQClusterReconciler rabbitClusterReconciler(
-            final RabbitMQClusterFactory clusterFactory,
-            final RabbitMQResourceController controller,
-            final SecretsController secretsController,
-            final ServicesController servicesController,
-            final StatefulSetController statefulSetController,
-            final PodDisruptionBudgetController podDisruptionBudgetController,
-            final PersistentVolumeClaimController persistentVolumeClaimController,
-            final ShovelReconciler shovelReconciler,
-            final ClusterUsersReconciler usersReconciler
-    ) {
-        return new RabbitMQClusterReconciler(clusterFactory, controller, secretsController, servicesController, statefulSetController, podDisruptionBudgetController, persistentVolumeClaimController, shovelReconciler, usersReconciler);
-    }
-
-    @Bean
-    public ShovelReconciler shovelReconciler(
-            final RabbitMQApiClient apiClient,
-            final SecretsController secretsController
-    ) {
-        return new ShovelReconciler(apiClient, secretsController);
-    }
-
-    @Bean
-    public NetworkPartitionWatcher networkPartitionWatcher(
-            final NetworkPartitionReconciler partitionReconciler,
-            final NetworkPartitionResourceController controller,
-            final ClusterReconciliationOrchestrator orchestrator
-    ) {
-        return new NetworkPartitionWatcher(partitionReconciler, controller, orchestrator);
-    }
-
-    @Bean
-    public ClusterUsersReconciler rabbitMQUserReconciler(
-            final SecretsController secretsController,
-            final RabbitMQApiClient apiClient,
-            final RabbitMQPasswordConverter passwordConverter
-    ) {
-        return new ClusterUsersReconciler(secretsController, apiClient, passwordConverter);
-    }
-
-    @Bean
-    public RabbitMQPasswordConverter passwordConverter() throws NoSuchAlgorithmException {
-        return new RabbitMQPasswordConverter(new Random(), MessageDigest.getInstance("SHA-256"), Base64.getEncoder(), Base64.getDecoder());
-    }
-
-    @Bean
-    public ClusterReconciliationOrchestrator clusterReconciliationOrchestrator(
-            final ClusterAwareExecutor executor
-    ) {
-        return new ClusterReconciliationOrchestrator(executor);
     }
 
     @Bean
